requirements.txt
python-twitter>=0
selenium==3.141.0
sentry-sdk>=0
